+++ crates/client/rpc/src/lib.rs
@@ -14,6 +14,7 @@ use std::sync::Arc;
 use dc_db::mapping_db::MappingDb;
 use dc_db::DeoxysBackend;
 use dp_block::{DeoxysBlock, DeoxysBlockInfo, DeoxysBlockInner};
+use dp_convert::felt_wrapper::FeltWrapper;
 use errors::StarknetRpcApiError;
 use jsonrpsee::core::RpcResult;
 use jsonrpsee::proc_macros::rpc;
@@ -192,7 +193,7 @@ pub trait StarknetTraceRpcApi {
 
 #[derive(Clone)]
 pub struct ChainConfig {
-    pub chain_id: Felt,
+    pub chain_id: starknet_types_core::felt::Felt,
     pub feeder_gateway: Url,
     pub gateway: Url,
 }
@@ -213,7 +214,7 @@ impl Starknet {
         SequencerGatewayProvider::new(
             self.chain_config.feeder_gateway.clone(),
             self.chain_config.gateway.clone(),
-            self.chain_config.chain_id.0,
+            self.chain_config.chain_id.into_field_element(),
         )
     }
 
@@ -254,7 +255,8 @@ impl Starknet {
     }
 
     fn chain_id(&self) -> RpcResult<Felt> {
-        Ok(self.chain_config.chain_id)
+        let chain_id_as_wrong_felt = Felt(self.chain_config.chain_id.into_field_element());
+        Ok(chain_id_as_wrong_felt)
     }
 
     pub fn current_block_number(&self) -> RpcResult<u64> {