+++ crates/client/rpc/src/methods/trace/trace_block_transactions.rs
@@ -1,4 +1,4 @@
-use dp_felt::FeltWrapper;
+use dp_convert::felt_wrapper::FeltWrapper;
 use dp_transactions::TxType;
 use jsonrpsee::core::RpcResult;
 use starknet_api::transaction::Transaction;