+++ crates/client/sync/src/l2.rs
@@ -14,10 +14,9 @@ use dc_telemetry::{TelemetryHandle, VerbosityLevel};
 use dp_block::Header;
 use dp_block::{BlockId, BlockTag, DeoxysBlock};
 use dp_convert::core_felt::CoreFelt;
-use dp_felt::FeltWrapper;
+use dp_convert::felt_wrapper::FeltWrapper;
 use futures::{stream, StreamExt};
 use num_traits::FromPrimitive;
-use starknet_api::hash::StarkFelt;
 use starknet_api::transaction::TransactionHash;
 use starknet_core::types::StateUpdate;
 use starknet_providers::sequencer::models::StateUpdateWithBlock;
@@ -198,7 +197,7 @@ pub struct L2ConvertedBlockAndUpdates {
 async fn l2_block_conversion_task(
     updates_receiver: mpsc::Receiver<L2BlockAndUpdates>,
     output: mpsc::Sender<L2ConvertedBlockAndUpdates>,
-    chain_id: StarkFelt,
+    chain_id: Felt,
 ) -> anyhow::Result<()> {
     // Items of this stream are futures that resolve to blocks, which becomes a regular stream of blocks
     // using futures buffered.
@@ -237,7 +236,7 @@ async fn l2_pending_block_task(
     backend: Arc<DeoxysBackend>,
     sync_finished_cb: oneshot::Receiver<()>,
     provider: Arc<SequencerGatewayProvider>,
-    chain_id: StarkFelt,
+    chain_id: Felt,
 ) -> anyhow::Result<()> {
     let backend = &backend;
 
@@ -306,7 +305,7 @@ pub async fn sync(
     config: L2SyncConfig,
     block_metrics: BlockMetrics,
     starting_block: u64,
-    chain_id: StarkFelt,
+    chain_id: Felt,
     telemetry: TelemetryHandle,
 ) -> anyhow::Result<()> {
     let (fetch_stream_sender, fetch_stream_receiver) = mpsc::channel(30);