+++ crates/primitives/transactions/src/to_starknet_core_transaction.rs
@@ -1,7 +1,7 @@
 use std::vec::Vec;
 
+use dp_convert::felt_wrapper::FeltWrapper;
 /// Here we transform starknet-api transactions into starknet-core trasnactions
-use dp_felt::Felt252Wrapper;
 use starknet_api::transaction::{
     DeclareTransaction, DeclareTransactionV0V1, DeclareTransactionV2, DeclareTransactionV3, DeployAccountTransaction,
     DeployAccountTransactionV1, DeployAccountTransactionV3, InvokeTransaction, InvokeTransactionV0,
@@ -22,14 +22,10 @@ pub fn to_starknet_core_tx(tx: &Transaction, transaction_hash: FieldElement) -> 
                     sender_address,
                 }) => starknet_core::types::DeclareTransaction::V0(starknet_core::types::DeclareTransactionV0 {
                     transaction_hash,
-                    max_fee: Felt252Wrapper::from(max_fee.0).into(),
-                    signature: signature
-                        .0
-                        .iter()
-                        .map(|x| Felt252Wrapper::from(*x).into())
-                        .collect::<Vec<FieldElement>>(),
-                    class_hash: Felt252Wrapper::from(class_hash.0).into(),
-                    sender_address: Felt252Wrapper::from(sender_address.0).into(),
+                    max_fee: FieldElement::from(max_fee.0),
+                    signature: signature.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
+                    class_hash: class_hash.0.into_field_element(),
+                    sender_address: sender_address.0.into_field_element(),
                 }),
                 DeclareTransaction::V1(DeclareTransactionV0V1 {
                     max_fee,
@@ -40,15 +36,11 @@ pub fn to_starknet_core_tx(tx: &Transaction, transaction_hash: FieldElement) -> 
                     ..
                 }) => starknet_core::types::DeclareTransaction::V1(starknet_core::types::DeclareTransactionV1 {
                     transaction_hash,
-                    max_fee: Felt252Wrapper::from(max_fee.0).into(),
-                    signature: signature
-                        .0
-                        .iter()
-                        .map(|x| Felt252Wrapper::from(*x).into())
-                        .collect::<Vec<FieldElement>>(),
-                    nonce: Felt252Wrapper::from(nonce.0).into(),
-                    class_hash: Felt252Wrapper::from(class_hash.0).into(),
-                    sender_address: Felt252Wrapper::from(sender_address.0).into(),
+                    max_fee: FieldElement::from(max_fee.0),
+                    signature: signature.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
+                    nonce: nonce.into_field_element(),
+                    class_hash: class_hash.into_field_element(),
+                    sender_address: sender_address.into_field_element(),
                 }),
                 DeclareTransaction::V2(DeclareTransactionV2 {
                     max_fee,
@@ -60,16 +52,12 @@ pub fn to_starknet_core_tx(tx: &Transaction, transaction_hash: FieldElement) -> 
                     ..
                 }) => starknet_core::types::DeclareTransaction::V2(starknet_core::types::DeclareTransactionV2 {
                     transaction_hash,
-                    max_fee: Felt252Wrapper::from(max_fee.0).into(),
-                    signature: signature
-                        .0
-                        .iter()
-                        .map(|x| Felt252Wrapper::from(*x).into())
-                        .collect::<Vec<FieldElement>>(),
-                    nonce: Felt252Wrapper::from(nonce.0).into(),
-                    class_hash: Felt252Wrapper::from(class_hash.0).into(),
-                    sender_address: Felt252Wrapper::from(sender_address.0).into(),
-                    compiled_class_hash: Felt252Wrapper::from(compiled_class_hash.0).into(),
+                    max_fee: FieldElement::from(max_fee.0),
+                    signature: signature.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
+                    nonce: nonce.into_field_element(),
+                    class_hash: class_hash.into_field_element(),
+                    sender_address: sender_address.into_field_element(),
+                    compiled_class_hash: compiled_class_hash.0.into_field_element(),
                 }),
                 DeclareTransaction::V3(DeclareTransactionV3 {
                     resource_bounds,
@@ -87,26 +75,22 @@ pub fn to_starknet_core_tx(tx: &Transaction, transaction_hash: FieldElement) -> 
                     transaction_hash,
                     resource_bounds: api_resources_to_core_ressources(resource_bounds),
                     tip: tip.0,
-                    signature: signature
-                        .0
-                        .iter()
-                        .map(|x| Felt252Wrapper::from(*x).into())
-                        .collect::<Vec<FieldElement>>(),
-                    nonce: Felt252Wrapper::from(nonce.0).into(),
-                    class_hash: Felt252Wrapper::from(class_hash.0).into(),
-                    compiled_class_hash: Felt252Wrapper::from(compiled_class_hash.0).into(),
-                    sender_address: Felt252Wrapper::from(sender_address.0).into(),
+                    signature: signature.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
+                    nonce: nonce.into_field_element(),
+                    class_hash: class_hash.into_field_element(),
+                    compiled_class_hash: compiled_class_hash.0.into_field_element(),
+                    sender_address: sender_address.into_field_element(),
                     nonce_data_availability_mode: api_da_to_core_da(nonce_data_availability_mode).unwrap(),
                     fee_data_availability_mode: api_da_to_core_da(fee_data_availability_mode).unwrap(),
                     paymaster_data: paymaster_data
                         .0
                         .iter()
-                        .map(|x| Felt252Wrapper::from(*x).into())
+                        .map(|x| x.into_field_element())
                         .collect::<Vec<FieldElement>>(),
                     account_deployment_data: account_deployment_data
                         .0
                         .iter()
-                        .map(|x| Felt252Wrapper::from(*x).into())
+                        .map(|x| x.into_field_element())
                         .collect::<Vec<FieldElement>>(),
                 }),
             };
@@ -126,20 +110,16 @@ pub fn to_starknet_core_tx(tx: &Transaction, transaction_hash: FieldElement) -> 
                 }) => starknet_core::types::DeployAccountTransaction::V1(
                     starknet_core::types::DeployAccountTransactionV1 {
                         transaction_hash,
-                        max_fee: Felt252Wrapper::from(max_fee.0).into(),
-                        signature: signature
-                            .0
-                            .iter()
-                            .map(|x| Felt252Wrapper::from(*x).into())
-                            .collect::<Vec<FieldElement>>(),
-                        nonce: Felt252Wrapper::from(nonce.0).into(),
-                        contract_address_salt: Felt252Wrapper::from(contract_address_salt.0).into(),
+                        max_fee: FieldElement::from(max_fee.0),
+                        signature: signature.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
+                        nonce: nonce.into_field_element(),
+                        contract_address_salt: contract_address_salt.0.into_field_element(),
                         constructor_calldata: constructor_calldata
                             .0
                             .iter()
-                            .map(|x| Felt252Wrapper::from(*x).into())
+                            .map(|x| x.into_field_element())
                             .collect::<Vec<FieldElement>>(),
-                        class_hash: Felt252Wrapper::from(class_hash.0).into(),
+                        class_hash: class_hash.into_field_element(),
                     },
                 ),
                 DeployAccountTransaction::V3(DeployAccountTransactionV3 {
@@ -158,25 +138,21 @@ pub fn to_starknet_core_tx(tx: &Transaction, transaction_hash: FieldElement) -> 
                         transaction_hash,
                         resource_bounds: api_resources_to_core_ressources(resource_bounds),
                         tip: tip.0,
-                        signature: signature
-                            .0
-                            .iter()
-                            .map(|x| Felt252Wrapper::from(*x).into())
-                            .collect::<Vec<FieldElement>>(),
-                        nonce: Felt252Wrapper::from(nonce.0).into(),
-                        class_hash: Felt252Wrapper::from(class_hash.0).into(),
-                        contract_address_salt: Felt252Wrapper::from(contract_address_salt.0).into(),
+                        signature: signature.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
+                        nonce: nonce.into_field_element(),
+                        class_hash: class_hash.into_field_element(),
+                        contract_address_salt: contract_address_salt.0.into_field_element(),
                         constructor_calldata: constructor_calldata
                             .0
                             .iter()
-                            .map(|x| Felt252Wrapper::from(*x).into())
+                            .map(|x| x.into_field_element())
                             .collect::<Vec<FieldElement>>(),
                         nonce_data_availability_mode: api_da_to_core_da(nonce_data_availability_mode).unwrap(),
                         fee_data_availability_mode: api_da_to_core_da(fee_data_availability_mode).unwrap(),
                         paymaster_data: paymaster_data
                             .0
                             .iter()
-                            .map(|x| Felt252Wrapper::from(*x).into())
+                            .map(|x| x.into_field_element())
                             .collect::<Vec<FieldElement>>(),
                     },
                 ),
@@ -187,15 +163,15 @@ pub fn to_starknet_core_tx(tx: &Transaction, transaction_hash: FieldElement) -> 
         Transaction::Deploy(tx) => {
             let tx = starknet_core::types::DeployTransaction {
                 transaction_hash,
-                contract_address_salt: Felt252Wrapper::from(tx.contract_address_salt.0).into(),
+                contract_address_salt: tx.contract_address_salt.0.into_field_element(),
                 constructor_calldata: tx
                     .constructor_calldata
                     .0
                     .iter()
-                    .map(|x| Felt252Wrapper::from(*x).into())
+                    .map(|x| x.into_field_element())
                     .collect::<Vec<FieldElement>>(),
-                class_hash: Felt252Wrapper::from(tx.class_hash.0).into(),
-                version: Felt252Wrapper::ZERO.into(),
+                class_hash: tx.class_hash.into_field_element(),
+                version: FieldElement::ZERO,
             };
 
             starknet_core::types::Transaction::Deploy(tx)
@@ -210,15 +186,11 @@ pub fn to_starknet_core_tx(tx: &Transaction, transaction_hash: FieldElement) -> 
                     calldata,
                 }) => starknet_core::types::InvokeTransaction::V0(starknet_core::types::InvokeTransactionV0 {
                     transaction_hash,
-                    max_fee: Felt252Wrapper::from(max_fee.0).into(),
-                    signature: signature
-                        .0
-                        .iter()
-                        .map(|x| Felt252Wrapper::from(*x).into())
-                        .collect::<Vec<FieldElement>>(),
-                    contract_address: (*contract_address.0.key()).into(),
-                    entry_point_selector: Felt252Wrapper::from(entry_point_selector.0).into(),
-                    calldata: calldata.0.iter().map(|x| Felt252Wrapper::from(*x).into()).collect::<Vec<FieldElement>>(),
+                    max_fee: FieldElement::from(max_fee.0),
+                    signature: signature.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
+                    contract_address: contract_address.into_field_element(),
+                    entry_point_selector: entry_point_selector.0.into_field_element(),
+                    calldata: calldata.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
                 }),
                 InvokeTransaction::V1(InvokeTransactionV1 {
                     max_fee,
@@ -229,15 +201,11 @@ pub fn to_starknet_core_tx(tx: &Transaction, transaction_hash: FieldElement) -> 
                     ..
                 }) => starknet_core::types::InvokeTransaction::V1(starknet_core::types::InvokeTransactionV1 {
                     transaction_hash,
-                    max_fee: Felt252Wrapper::from(max_fee.0).into(),
-                    signature: signature
-                        .0
-                        .iter()
-                        .map(|x| Felt252Wrapper::from(*x).into())
-                        .collect::<Vec<FieldElement>>(),
-                    nonce: Felt252Wrapper::from(nonce.0).into(),
-                    sender_address: Felt252Wrapper::from(sender_address.0).into(),
-                    calldata: calldata.0.iter().map(|x| Felt252Wrapper::from(*x).into()).collect::<Vec<FieldElement>>(),
+                    max_fee: FieldElement::from(max_fee.0),
+                    signature: signature.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
+                    nonce: nonce.into_field_element(),
+                    sender_address: sender_address.into_field_element(),
+                    calldata: calldata.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
                 }),
                 InvokeTransaction::V3(InvokeTransactionV3 {
                     resource_bounds,
@@ -254,25 +222,21 @@ pub fn to_starknet_core_tx(tx: &Transaction, transaction_hash: FieldElement) -> 
                     transaction_hash,
                     resource_bounds: api_resources_to_core_ressources(resource_bounds),
                     tip: tip.0,
-                    signature: signature
-                        .0
-                        .iter()
-                        .map(|x| Felt252Wrapper::from(*x).into())
-                        .collect::<Vec<FieldElement>>(),
-                    nonce: Felt252Wrapper::from(nonce.0).into(),
-                    sender_address: Felt252Wrapper::from(sender_address.0).into(),
-                    calldata: calldata.0.iter().map(|x| Felt252Wrapper::from(*x).into()).collect::<Vec<FieldElement>>(),
+                    signature: signature.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
+                    nonce: nonce.into_field_element(),
+                    sender_address: sender_address.into_field_element(),
+                    calldata: calldata.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
                     nonce_data_availability_mode: api_da_to_core_da(nonce_data_availability_mode).unwrap(),
                     fee_data_availability_mode: api_da_to_core_da(fee_data_availability_mode).unwrap(),
                     paymaster_data: paymaster_data
                         .0
                         .iter()
-                        .map(|x| Felt252Wrapper::from(*x).into())
+                        .map(|x| x.into_field_element())
                         .collect::<Vec<FieldElement>>(),
                     account_deployment_data: account_deployment_data
                         .0
                         .iter()
-                        .map(|x| Felt252Wrapper::from(*x).into())
+                        .map(|x| x.into_field_element())
                         .collect::<Vec<FieldElement>>(),
                 }),
             };
@@ -283,10 +247,10 @@ pub fn to_starknet_core_tx(tx: &Transaction, transaction_hash: FieldElement) -> 
             let tx = starknet_core::types::L1HandlerTransaction {
                 transaction_hash,
                 version: FieldElement::ZERO,
-                nonce: u64::try_from(Felt252Wrapper::from(tx.nonce.0)).unwrap(),
-                contract_address: Felt252Wrapper::from(tx.contract_address).into(),
-                entry_point_selector: Felt252Wrapper::from(tx.entry_point_selector).into(),
-                calldata: tx.calldata.0.iter().map(|x| Felt252Wrapper::from(*x).into()).collect::<Vec<FieldElement>>(),
+                nonce: u64::try_from(tx.nonce.0).unwrap(),
+                contract_address: tx.contract_address.into_field_element(),
+                entry_point_selector: tx.entry_point_selector.0.into_field_element(),
+                calldata: tx.calldata.0.iter().map(|x| x.into_field_element()).collect::<Vec<FieldElement>>(),
             };
 
             starknet_core::types::Transaction::L1Handler(tx)